requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3

# LangGraph and LangChain
langgraph>=0.0.27
langchain>=0.0.312
langchain-core>=0.0.11
langchain-openai>=0.0.2
openai>=1.3.0
langchain-mcp-adapters>=0.0.1

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.28.0
psycopg2-binary>=2.9.9

# Cloud storage
boto3>=1.28.63

# Authentication and security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Utilities
uuid>=1.30
httpx>=0.24.1
tenacity>=8.2.3
pytest>=7.4.3
